feat(api): reorganized the code for clarity
refs. metriport/metriport-internal#1350
