+++ packages/api/src/external/carequality/patient.ts
@@ -1,5 +1,5 @@
 import { Patient, PatientExternalData } from "@metriport/core/domain/patient";
-import { getOrgs } from "@metriport/core/external/commonwell/cq-bridge/get-orgs";
+
 import { toFHIR } from "@metriport/core/external/fhir/patient/index";
 import { MedicalDataSource } from "@metriport/core/external/index";
 import { out } from "@metriport/core/util/log";
@@ -12,16 +12,14 @@ import { getOrganizationOrFail } from "../../command/medical/organization/get-or
 import { isCQDirectEnabledForCx } from "../aws/appConfig";
 import { makeIheGatewayAPIForPatientDiscovery } from "../ihe-gateway/api";
 import { makeOutboundResultPoller } from "../ihe-gateway/outbound-result-poller-factory";
-import { getOrganizationsWithXCPD } from "./command/cq-directory/cq-gateways";
+import { getGatewaysAndNonGateways } from "./command/cq-directory/cq-gateways";
 import {
-  CQOrgBasicDetails,
   filterCQOrgsToSearch,
   searchCQDirectoriesAroundPatientAddresses,
-  toBasicOrgAttributes,
 } from "./command/cq-directory/search-cq-directory";
+import { sortCQOrganizationsByPrio } from "./command/cq-directory/sort-cq-organizations";
 import { deleteCQPatientData } from "./command/cq-patient-data/delete-cq-data";
 import { createOutboundPatientDiscoveryReq } from "./create-outbound-patient-discovery-req";
-import { CQDirectoryEntryModel } from "./models/cq-directory";
 import { cqOrgsToXCPDGateways, generateIdsForGateways } from "./organization-conversion";
 import { PatientDataCarequality } from "./patient-shared";
 
@@ -30,7 +28,6 @@ dayjs.extend(duration);
 const context = "cq.patient.discover";
 const iheGateway = makeIheGatewayAPIForPatientDiscovery();
 const resultPoller = makeOutboundResultPoller();
-const cqOrgsHydrated = getOrgs();
 
 export async function discover(patient: Patient, facilityNPI: string): Promise<void> {
   const baseLogMessage = `CQ PD - patientId ${patient.id}`;
@@ -92,12 +89,12 @@ async function prepareForPatientDiscovery(
   const [organization, nearbyOrgs, allOrgs] = await Promise.all([
     getOrganizationOrFail({ cxId }),
     searchCQDirectoriesAroundPatientAddresses({ patient }),
-    getOrganizationsWithXCPD(),
+    getGatewaysAndNonGateways(),
   ]);
-  const [gateways, orgs] = splitGatewaysAndNonGateways(allOrgs);
-  const sortedGateways = sortCQOrganizationsByPrio(gateways);
+
+  const sortedGateways = sortCQOrganizationsByPrio(allOrgs.gateways);
   const sortedNearbyOrgs = sortCQOrganizationsByPrio(nearbyOrgs);
-  const sortedOrgs = sortCQOrganizationsByPrio(orgs);
+  const sortedOrgs = sortCQOrganizationsByPrio(allOrgs.nonGateways);
 
   const orgsToSearch = filterCQOrgsToSearch([
     ...sortedGateways,
@@ -116,37 +113,4 @@ async function prepareForPatientDiscovery(
     orgOid: organization.oid,
   });
   return pdRequest;
-}
-
-function splitGatewaysAndNonGateways(
-  cqOrgs: CQDirectoryEntryModel[]
-): [CQDirectoryEntryModel[], CQDirectoryEntryModel[]] {
-  return [cqOrgs.filter(o => o.gateway), cqOrgs.filter(o => !o.gateway)];
-}
-
-function sortCQOrganizationsByPrio(cqOrgs: CQDirectoryEntryModel[]): CQOrgBasicDetails[] {
-  const orgsWithPrio = mapCQOrganizationPriorities(cqOrgs);
-  const sortedOrgs = sortByPrio(orgsWithPrio);
-  const sortedOrgBasics = sortedOrgs.map(toBasicOrgAttributes);
-  return sortedOrgBasics;
-}
-
-type CQOrganizationWithPrio = CQDirectoryEntryModel & { prio: string | undefined };
-
-function mapCQOrganizationPriorities(cqOrgs: CQDirectoryEntryModel[]): CQOrganizationWithPrio[] {
-  const res = cqOrgs.map(org => {
-    const matchingOrg = cqOrgsHydrated.find(o => o.id === org.id);
-    return {
-      ...org.dataValues,
-      prio: matchingOrg ? matchingOrg.prio : undefined,
-    } as CQOrganizationWithPrio;
-  });
-  return res;
-}
-
-function sortByPrio(orgs: CQOrganizationWithPrio[]): CQOrganizationWithPrio[] {
-  const high = orgs.filter(o => o.prio === "high");
-  const medium = orgs.filter(o => o.prio === "medium");
-  const low = orgs.filter(o => o.prio === "low" || !o.prio);
-  return [...high, ...medium, ...low];
 }
